requirements.txt
fastapi==0.70.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.2
python-dotenv==0.19.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
sqlalchemy==1.4.27
uvicorn==0.15.0
